[REFACTOR] Renamed analyser.py to sentiment_analyser.py
