+++ main.py
@@ -1,4 +1,4 @@
-import analyser
+import sentiment_analyser
 import drawer
 import reader
 import pathlib
@@ -8,7 +8,7 @@ def main(text_path: str, save_image: bool = True):
     if not pathlib.Path(text_path).is_file():
         print(text_path + " is not a valid path. Please enter a valid path and try again")
         return
-    analyser.initialisation()
+    sentiment_analyser.initialisation()
     drawer.initialisation()
     lines = reader.read_lines(text_path)
     for line in lines:
